Add EntryStatusMessage component
