+++ gym_managementservice_frontend/src/pages/EntryStatusPage.module.css
@@ -8,8 +8,3 @@
     text-align: center;
 }
 
-.chill {
-    font-size: 8vw;
-    font-weight: bold;
-    line-height: 1.1;
-}
